+++ tests/Unit/DSC_UserAccountControl.Tests.ps1
@@ -108,15 +108,9 @@ Describe 'UserAccountControl\Get-TargetResource' -Tag 'Get' {
 
 Describe 'UserAccountControl\Set-TargetResource' -Tag 'Set' {
     BeforeAll {
-        Mock -CommandName Assert-BoundParameter
+        Mock -CommandName Assert-BoundParameter -RemoveParameterType @('RequiredBehavior')
         Mock -CommandName Set-ItemProperty
         Mock -CommandName Set-UserAccountControlToNotificationLevel
-
-        InModuleScope -ScriptBlock {
-            function script:Assert-BoundParameter {
-                param ()
-            }
-        }
     }
 
     BeforeEach {