+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - `UserAccountControl`
-  - Add dummy `Assert-BoundParameter` function to please Pester.
+  - Tests remove parameter type on `Assert-BoundParameter` mocks.
 
 ## [10.0.0] - 2025-01-25